requirements/base.txt
# General
# ------------------------------------------------------------------------------
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
inflection==0.5.1
integer==1.0.3
Markdown==3.7
packaging==24.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
whitenoise==6.9.0

# Django
# ------------------------------------------------------------------------------
Django==4.2.7
django-filter==23.5
django-test-plus==2.2.4
django-environ==0.12.0

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.15.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.4.0